updated game
Included new feature Times Tables where user can choose an number and an upper limit (multiple of that number) and then answer problems that are just multiplication practice for that number. This did introduce one minor bug that I need to iron out. Also added a new modal to allow the entry of the Times Table data.